运输管理发运调用发运回传接口

+++ src/main/java/com/aiqin/bms/scmp/api/purchase/domain/pojo/transport/TransportOrders.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -28,7 +29,7 @@ public class TransportOrders extends CommonBean {
 
     @ApiModelProperty("订单金额")
 //    @JsonProperty("order_amount")
-    private Long orderAmount;
+    private BigDecimal orderAmount;
 
     @ApiModelProperty("订单状态")
 //    @JsonProperty("status")
